YTWapi/requirements.txt
django==3.1
celery==4.4.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
PyMySQL==0.10.1
mysqlclient==2.0.1
redis==3.5.3
requests==2.24.0
